Add bottom navigation bar with 4 screens
