--- androidApp/src/main/java/com/helios/kmptranslator/android/TranslateApp.kt
@@ -1,8 +1,0 @@
-package com.helios.kmptranslator.android
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class TranslateApp: Application() {
-}
