+++ androidApp/src/main/java/com/helios/kmptranslator/android/TranslateApplication.kt
@@ -1,0 +1,8 @@
+package com.helios.kmptranslator.android
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class TranslateApplication: Application() {
+}
