+++ androidApp/src/main/java/com/helios/kmptranslator/android/app/navigation/NavHost.kt
@@ -1,0 +1,25 @@
+package com.helios.kmptranslator.android.app.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import com.helios.kmptranslator.android.scantranslate.navigation.scanTranslateScreen
+import com.helios.kmptranslator.android.settings.navigation.settingScreen
+import com.helios.kmptranslator.android.texttranslate.navigation.TEXT_TRANSLATE_ROUTE
+import com.helios.kmptranslator.android.texttranslate.navigation.textTranslateScreen
+import com.helios.kmptranslator.android.voicetranslate.navigation.voiceTranslateScreen
+
+@Composable
+fun AppNavHost(
+    navController: NavHostController,
+    modifier: Modifier = Modifier,
+    startDestination: String = TEXT_TRANSLATE_ROUTE
+) {
+    NavHost(navController = navController, startDestination = startDestination, modifier = modifier) {
+        textTranslateScreen()
+        voiceTranslateScreen()
+        scanTranslateScreen()
+        settingScreen()
+    }
+}
