+++ androidApp/src/main/java/com/helios/kmptranslator/android/scantranslate/ScanTranslateScreen.kt
@@ -1,0 +1,14 @@
+package com.helios.kmptranslator.android.scantranslate
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun ScanTranslateScreen() {
+    Box(modifier = Modifier.fillMaxSize()) {
+        Text(text = "Scan and Translate")
+    }
+}
