+++ androidApp/src/main/java/com/helios/kmptranslator/android/settings/SettingScreen.kt
@@ -1,0 +1,14 @@
+package com.helios.kmptranslator.android.settings
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun SettingScreen() {
+    Box(modifier = Modifier.fillMaxSize()) {
+        Text(text = "Settings")
+    }
+}
