+++ androidApp/src/main/java/com/helios/kmptranslator/android/texttranslate/components/LanguageDropDown.kt
@@ -1,0 +1,74 @@
+package com.helios.kmptranslator.android.texttranslate.components
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowDropDown
+import androidx.compose.material.icons.filled.ArrowDropUp
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.helios.kmptranslator.android.R
+import com.helios.kmptranslator.android.core.theme.LightBlue
+import com.helios.kmptranslator.core.presentation.UiLanguage
+
+@Composable
+fun LanguageDropDown(
+    language: UiLanguage,
+    isOpen: Boolean,
+    onClick: () -> Unit,
+    onDismiss: () -> Unit,
+    onSelectLanguage: (UiLanguage) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Box(modifier = modifier) {
+        if (isOpen) {
+            DropdownMenu(
+                expanded = true,
+                onDismissRequest = onDismiss
+            ) {
+                UiLanguage.allLanguages.forEach { language ->
+                    LanguageDropDownItem(
+                        language = language,
+                        onClick = { onSelectLanguage(language) },
+                        modifier = modifier
+                    )
+                }
+            }
+        }
+
+        Row(
+            modifier = Modifier
+                .clickable { onClick() }
+                .padding(16.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            AsyncImage(
+                model = language.drawableRes,
+                contentDescription = language.language.langName,
+                modifier = Modifier.size(30.dp)
+            )
+            Spacer(modifier = Modifier.width(16.dp))
+            Text(text = language.language.langName, color = LightBlue)
+            Icon(
+                imageVector = if (isOpen) Icons.Default.ArrowDropUp else Icons.Default.ArrowDropDown,
+                contentDescription = if (isOpen) {
+                    stringResource(id = R.string.close_language_picker)
+                } else {
+                    stringResource(id = R.string.open)
+                }
+            )
+        }
+    }
+}
