+++ androidApp/src/main/java/com/helios/kmptranslator/android/texttranslate/components/SwapLanguagesButton.kt
@@ -1,0 +1,36 @@
+package com.helios.kmptranslator.android.texttranslate.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import com.helios.kmptranslator.android.R
+
+@Composable
+fun SwapLanguagesButton(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+
+    IconButton(
+        onClick = onClick,
+        modifier
+            .clip(CircleShape)
+            .background(MaterialTheme.colorScheme.primary)
+    ) {
+        Icon(
+            imageVector = ImageVector.vectorResource(id = R.drawable.swap_languages),
+            contentDescription = stringResource(
+                id = R.string.swap_languages
+            ),
+            tint = MaterialTheme.colorScheme.onPrimary
+        )
+    }
+}
