+++ androidApp/src/main/java/com/helios/kmptranslator/android/texttranslate/components/TranslateTextField.kt
@@ -1,0 +1,211 @@
+package com.helios.kmptranslator.android.texttranslate.components
+
+import androidx.compose.animation.AnimatedContent
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Close
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.graphics.SolidColor
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.helios.kmptranslator.android.R
+import com.helios.kmptranslator.android.core.theme.LightBlue
+import com.helios.kmptranslator.core.presentation.UiLanguage
+
+@Composable
+fun TranslateTextField(
+    fromText: String,
+    toText: String?,
+    isTranslating: Boolean,
+    fromLanguage: UiLanguage,
+    toLanguage: UiLanguage,
+    onTranslateClick: () -> Unit,
+    onTextChanged: (String) -> Unit,
+    onCopyClick: (String) -> Unit,
+    onCloseClick: () -> Unit,
+    onSpeakerClick: () -> Unit,
+    onTextFieldClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Box(
+        modifier = modifier
+            .shadow(elevation = 5.dp, shape = RoundedCornerShape(20.dp))
+            .clip(RoundedCornerShape(20.dp))
+            .gradientSurface()
+            .clickable { onTextFieldClick() }
+            .padding(16.dp)
+    ) {
+        AnimatedContent(
+            targetState = toText,
+            label = "shouldShowIdleState"
+        ) { toText ->
+            if (toText == null || isTranslating) {
+                IdleTranslateTextField(
+                    fromText = fromText,
+                    isTranslating = isTranslating,
+                    onTextChanged = onTextChanged,
+                    onTranslateClick = onTranslateClick,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .aspectRatio(2f)
+                )
+            } else {
+                TranslatedTextField(
+                    fromText = fromText,
+                    toText = toText,
+                    fromLanguage = fromLanguage,
+                    toLanguage = toLanguage,
+                    onCopyClick = onCopyClick,
+                    onCloseClick = onCloseClick,
+                    onSpeakerClick = onSpeakerClick,
+                    modifier = Modifier.fillMaxWidth()
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun TranslatedTextField(
+    fromText: String,
+    toText: String,
+    fromLanguage: UiLanguage,
+    toLanguage: UiLanguage,
+    onCopyClick: (String) -> Unit,
+    onCloseClick: () -> Unit,
+    onSpeakerClick: () -> Unit,
+    modifier: Modifier
+) {
+
+    Column(modifier = modifier) {
+        LanguageDisplay(language = fromLanguage)
+        Spacer(modifier = Modifier.height(16.dp))
+        Text(text = fromText, color = MaterialTheme.colorScheme.onSurface)
+        Spacer(modifier = Modifier.height(16.dp))
+        Row(modifier = Modifier.align(Alignment.End)) {
+            IconButton(onClick = { onCopyClick(fromText) }) {
+                Icon(
+                    imageVector = ImageVector.vectorResource(id = R.drawable.copy),
+                    contentDescription = stringResource(id = R.string.copy),
+                    tint = LightBlue
+                )
+            }
+            IconButton(onClick = { onCloseClick() }) {
+                Icon(
+                    imageVector = Icons.Rounded.Close,
+                    contentDescription = stringResource(id = R.string.close_translated_text_field),
+                    tint = LightBlue
+                )
+            }
+        }
+        Spacer(modifier = Modifier.height(16.dp))
+        Divider()
+        Spacer(modifier = Modifier.height(16.dp))
+        LanguageDisplay(language = toLanguage)
+        Spacer(modifier = Modifier.height(16.dp))
+        Text(text = toText, color = MaterialTheme.colorScheme.onSurface)
+        Spacer(modifier = Modifier.height(16.dp))
+        Row(modifier = Modifier.align(Alignment.End)) {
+            IconButton(onClick = { onCopyClick(toText) }) {
+                Icon(
+                    imageVector = ImageVector.vectorResource(id = R.drawable.copy),
+                    contentDescription = stringResource(id = R.string.copy),
+                    tint = LightBlue
+                )
+            }
+            IconButton(onClick = { onSpeakerClick() }) {
+                Icon(
+                    imageVector = ImageVector.vectorResource(R.drawable.speaker),
+                    contentDescription = stringResource(id = R.string.play_loud),
+                    tint = LightBlue
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun IdleTranslateTextField(
+    fromText: String,
+    isTranslating: Boolean,
+    onTextChanged: (String) -> Unit,
+    onTranslateClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    var isFocused by remember {
+        mutableStateOf(false)
+    }
+
+    Box(modifier = modifier) {
+        BasicTextField(
+            value = fromText,
+            onValueChange = onTextChanged,
+            cursorBrush = SolidColor(MaterialTheme.colorScheme.primary),
+            modifier = Modifier
+                .fillMaxSize()
+                .onFocusChanged { isFocused = it.isFocused },
+            textStyle = TextStyle(color = MaterialTheme.colorScheme.onSurface)
+        )
+        if (fromText.isEmpty() && !isFocused) {
+            Text(
+                text = stringResource(id = R.string.enter_a_text_to_translate),
+                color = LightBlue
+            )
+        }
+        ProgressButton(
+            text = stringResource(R.string.translate),
+            inProgress = isTranslating,
+            onClick = onTranslateClick,
+            modifier = Modifier
+                .align(
+                    Alignment.BottomEnd
+                )
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun TranslateTextFieldPreview() {
+    TranslateTextField(
+        fromText = "",
+        toText = null,
+        isTranslating = false,
+        fromLanguage = UiLanguage.allLanguages[0],
+        toLanguage = UiLanguage.allLanguages[1],
+        onTranslateClick = { /*TODO*/ },
+        onTextChanged = {},
+        onCopyClick = {},
+        onCloseClick = { },
+        onSpeakerClick = { },
+        onTextFieldClick = { })
+}
