+++ androidApp/src/main/java/com/helios/kmptranslator/android/texttranslate/components/TranslationHistoryItem.kt
@@ -1,0 +1,71 @@
+package com.helios.kmptranslator.android.texttranslate.components
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import com.helios.kmptranslator.android.core.theme.LightBlue
+import com.helios.kmptranslator.core.presentation.UiHistoryItem
+
+@Composable
+fun TranslationHistoryItem(
+    historyItem: UiHistoryItem,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier
+            .shadow(
+                elevation = 5.dp,
+                shape = RoundedCornerShape(20.dp)
+            )
+            .clip(RoundedCornerShape(20.dp))
+            .gradientSurface()
+            .clickable(onClick = onClick)
+            .padding(16.dp)
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            SmallLanguageIcon(language = historyItem.fromLanguage)
+            Spacer(modifier = Modifier.width(16.dp))
+            Text(
+                text = historyItem.fromText,
+                style = MaterialTheme.typography.bodyMedium,
+                color = LightBlue,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+        }
+        Spacer(modifier = Modifier.height(24.dp))
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            SmallLanguageIcon(language = historyItem.toLanguage)
+            Spacer(modifier = Modifier.width(16.dp))
+            Text(
+                text = historyItem.toText,
+                style = MaterialTheme.typography.bodyLarge,
+                color = MaterialTheme.colorScheme.onSurface,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+        }
+    }
+}
