+++ androidApp/src/main/java/com/helios/kmptranslator/android/texttranslate/components/rememberTextToSpeech.kt
@@ -1,0 +1,24 @@
+package com.helios.kmptranslator.android.texttranslate.components
+
+import android.speech.tts.TextToSpeech
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.remember
+import androidx.compose.ui.platform.LocalContext
+
+@Composable
+fun rememberTextToSpeech(): TextToSpeech {
+    val context = LocalContext.current
+    val tts = remember {
+        TextToSpeech(context, null)
+    }
+
+    DisposableEffect(key1 = tts) {
+        onDispose {
+            tts.stop()
+            tts.shutdown()
+        }
+    }
+
+    return tts
+}
