--- androidApp/src/main/java/com/helios/kmptranslator/android/voicetotext/domain/AndroidVoiceToTextParser.kt
@@ -1,119 +1,0 @@
-package com.helios.kmptranslator.android.voicetotext.domain
-
-import android.app.Application
-import android.content.Intent
-import android.os.Bundle
-import android.speech.RecognitionListener
-import android.speech.RecognizerIntent
-import android.speech.SpeechRecognizer
-import com.helios.kmptranslator.android.R
-import com.helios.kmptranslator.core.util.CommonStateFlow
-import com.helios.kmptranslator.core.util.toCommonStateFlow
-import com.helios.kmptranslator.voicetotext.domain.VoiceToTextParser
-import com.helios.kmptranslator.voicetotext.domain.VoiceToTextParserState
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.update
-
-class AndroidVoiceToTextParser(
-    private val app: Application,
-) : VoiceToTextParser, RecognitionListener {
-
-    private val recognizer = SpeechRecognizer.createSpeechRecognizer(app)
-
-    private val _state = MutableStateFlow(VoiceToTextParserState())
-
-    override val state: CommonStateFlow<VoiceToTextParserState>
-        get() = _state.toCommonStateFlow()
-
-    override fun startListening(languageCode: String) {
-        _state.update { VoiceToTextParserState() }
-
-        if (SpeechRecognizer.isRecognitionAvailable(app)) {
-            _state.update {
-                it.copy(
-                    error = app.getString(R.string.recognition_not_available)
-                )
-            }
-            return
-        }
-
-        val intent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH).apply {
-            putExtra(
-                RecognizerIntent.EXTRA_LANGUAGE_MODEL,
-                RecognizerIntent.LANGUAGE_MODEL_FREE_FORM
-            )
-            putExtra(RecognizerIntent.EXTRA_LANGUAGE, languageCode)
-        }
-
-        recognizer.setRecognitionListener(this)
-        recognizer.startListening(intent)
-
-        _state.update {
-            it.copy(isSpeaking = true)
-        }
-    }
-
-    override fun stopListening() {
-        _state.update { VoiceToTextParserState() }
-        recognizer.stopListening()
-    }
-
-    override fun cancel() {
-        recognizer.cancel()
-    }
-
-    override fun reset() {
-        _state.value = VoiceToTextParserState()
-    }
-
-    override fun onReadyForSpeech(params: Bundle?) {
-        _state.update {
-            it.copy(
-                error = null
-            )
-        }
-    }
-
-    override fun onBeginningOfSpeech() = Unit
-
-    override fun onRmsChanged(rmsdB: Float) {
-        _state.update { state ->
-            state.copy(
-                powerRatio = rmsdB * (1f / (DECIBEL_RANGE))
-            )
-        }
-    }
-
-    override fun onBufferReceived(buffer: ByteArray?) = Unit
-
-    override fun onEndOfSpeech() {
-        _state.update { state ->
-            state.copy(
-                isSpeaking = false
-            )
-        }
-    }
-
-    override fun onError(error: Int) {
-        _state.update { it.copy(error = "Error: $error") }
-    }
-
-    override fun onResults(results: Bundle?) {
-        results
-            ?.getStringArrayList(SpeechRecognizer.RESULTS_RECOGNITION)
-            ?.getOrNull(0)
-            ?.let { text ->
-                _state.update { it.copy(result = text) }
-            }
-    }
-
-    override fun onPartialResults(partialResults: Bundle?) = Unit
-
-    override fun onEvent(eventType: Int, params: Bundle?) = Unit
-
-    companion object {
-        private const val MIN_DECIBEL = -2f
-        private const val MAX_DECIBEL = 12f
-        private const val DECIBEL_RANGE = MAX_DECIBEL - MIN_DECIBEL
-    }
-}
