+++ androidApp/src/main/res/values/strings.xml
@@ -23,4 +23,8 @@
     <string name="close">Close</string>
     <string name="start_talking">Start talking</string>
     <string name="record_voice">Voice record</string>
+    <string name="feature_text_translate_title">Typing</string>
+    <string name="feature_voice_translate_title">Voice</string>
+    <string name="feature_scan_translate_title">Scan</string>
+    <string name="settings">Settings</string>
 </resources>
